requirements.txt
Django==1.7
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
django-audit-log==0.5.1
django-celery==3.1.16
django-tastypie==0.12.1
kombu==3.0.26
lxml==3.4.4
oauthlib==0.7.2
python-dateutil==2.4.2
python-mimeparse==0.1.4
pytz==2015.4
redis==2.10.3
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
tweepy==3.3.0
wsgiref==0.1.2
